requirements.txt
streamlit>=1.30.0
openpyxl>=3.1.2
python-pptx>=0.6.23
pandas>=2.1.0
anthropic>=0.40.0
